requirements.txt
torch
torchvision
opencv-python
av
pandas
numpy
matplotlib
pyyaml
tqdm
